+++ statsmodels/regression/process_regression.py
